+++ distorch_2.py
@@ -843,406 +843,9 @@ def calculate_safetensor_vvram_allocation(model_patcher, virtual_vram_str):
     allocations_string = ";".join(allocation_parts)
     return allocations_string
 
-def override_class_with_distorch_safetensor_v2(cls):
-    """DisTorch 2.0 wrapper for safetensor models"""
-
-    class NodeOverrideDisTorchSafetensorV2(cls):
-        @classmethod
-        def INPUT_TYPES(s):
-            inputs = copy.deepcopy(cls.INPUT_TYPES())
-            devices = get_device_list()
-            compute_device = devices[1] if len(devices) > 1 else devices[0]
-
-            inputs["optional"] = inputs.get("optional", {})
-            inputs["optional"]["compute_device"] = (devices, {"default": compute_device})
-            inputs["optional"]["virtual_vram_gb"] = ("FLOAT", {"default": 4.0, "min": 0.0, "max": 128.0, "step": 0.1})
-            inputs["optional"]["donor_device"] = (devices, {"default": "cpu"})
-            inputs["optional"]["expert_mode_allocations"] = ("STRING", {"multiline": False, "default": ""})
-            inputs["optional"]["keep_loaded"] = ("BOOLEAN", {"default": True})
-            return inputs
-
-        CATEGORY = "multigpu/distorch_2"
-        FUNCTION = "override"
-        TITLE = f"{cls.TITLE if hasattr(cls, 'TITLE') else cls.__name__} (DisTorch2)"
-
-        @classmethod
-        def IS_CHANGED(s, *args, compute_device=None, virtual_vram_gb=4.0,
-                       donor_device="cpu", expert_mode_allocations="", keep_loaded=True, **kwargs):
-            settings_str = f"{compute_device}{virtual_vram_gb}{donor_device}{expert_mode_allocations}{keep_loaded}"
-            current_hash = hashlib.sha256(settings_str.encode()).hexdigest()
-
-            if not hasattr(cls, '_last_hash'):
-                cls._last_hash = current_hash
-                logger.mgpu_mm_log(f"IS_CHANGED first call: {current_hash[:8]}")
-            elif cls._last_hash != current_hash:
-                cls._last_hash = current_hash
-                logger.mgpu_mm_log(f"IS_CHANGED CHANGED: {current_hash[:8]} ← settings changed")
-            return current_hash
-
-        def override(self, *args, compute_device=None, virtual_vram_gb=4.0,
-                     donor_device="cpu", expert_mode_allocations="", keep_loaded=True, **kwargs):
-
-            unload_distorch_model = not keep_loaded
-
-            from . import set_current_device
-            if compute_device is not None:
-                set_current_device(compute_device)
-
-            # Register our patched ModelPatcher
-            register_patched_safetensor_modelpatcher()
-
-            # Build allocation string
-            vram_string = ""
-            if virtual_vram_gb > 0:
-                vram_string = f"{compute_device};{virtual_vram_gb};{donor_device}"
-            elif expert_mode_allocations:  # Only include compute device if there's an expert string
-                vram_string = compute_device
-
-            full_allocation = f"{expert_mode_allocations}#{vram_string}" if expert_mode_allocations or vram_string else ""
-            
-            fn = getattr(super(), cls.FUNCTION)
-            
-            # Load the model and get hash, then store allocation for future runs
-            out = fn(*args, **kwargs)
-            
-            model_to_check = None
-            if hasattr(out[0], 'model'):
-                model_to_check = out[0]
-            elif hasattr(out[0], 'patcher') and hasattr(out[0].patcher, 'model'):
-                model_to_check = out[0].patcher
-
-            if model_to_check:
-                model_hash = create_safetensor_model_hash(model_to_check, "override_store")
-                settings_str = f"{compute_device}{virtual_vram_gb}{donor_device}{expert_mode_allocations}"
-                settings_hash = hashlib.sha256(settings_str.encode()).hexdigest()
-                
-                # Store allocation for next run - this enables DisTorch for subsequent loads
-                safetensor_allocation_store[model_hash] = full_allocation
-                safetensor_settings_store[model_hash] = settings_hash
-                logger.debug(f"[MultiGPU DisTorch V2] Stored allocation for model {model_hash[:8]}: {full_allocation}")
-
-            logger.info(f"[MultiGPU DisTorch V2] Full allocation string: {full_allocation}")
-
-            logger.mgpu_mm_log(f"[FLAG_SET_START] Setting '_mgpu_unload_distorch_model' to: {unload_distorch_model} (keep_loaded={keep_loaded})")
-
-            # DIAGNOSTIC: Log full object chain at SET time
-            if hasattr(out[0], 'model'):
-                mp = out[0]  # This is the ModelPatcher
-                mp_id = id(mp)
-                inner_model = getattr(mp, 'model', None)
-                inner_model_id = id(inner_model) if inner_model else None
-                inner_model_name = type(inner_model).__name__ if inner_model else "None"
-                
-                # Format inner_model_id properly for f-string
-                inner_id_str = f"0x{inner_model_id:x}" if inner_model_id is not None else "None"
-                
-                logger.mgpu_mm_log(f"[OBJECT_CHAIN_SET] ModelPatcher: mp_id=0x{mp_id:x}, inner_model_id={inner_id_str}, inner_model_type={inner_model_name}")
-                
-                # FIX: Store flag on ModelPatcher itself (not inner model)
-                # This aligns with where it will be READ in model_management_mgpu.py
-                mp._mgpu_unload_distorch_model = unload_distorch_model
-                logger.mgpu_mm_log(f"[FLAG_SET_LOCATION] Set on ModelPatcher (mp_id=0x{mp_id:x}): mp._mgpu_unload_distorch_model = {unload_distorch_model}")
-                
-                # Also set on inner model for backwards compatibility during transition
-                if inner_model:
-                    inner_model._mgpu_unload_distorch_model = unload_distorch_model
-                    logger.mgpu_mm_log(f"[FLAG_SET_COMPAT] Also set on inner model (inner_model_id=0x{inner_model_id:x}) for compatibility")
-                    
-            elif hasattr(out[0], 'patcher') and hasattr(out[0].patcher, 'model'):
-                mp = out[0].patcher  # This is the ModelPatcher
-                mp_id = id(mp)
-                inner_model = getattr(mp, 'model', None)
-                inner_model_id = id(inner_model) if inner_model else None
-                inner_model_name = type(inner_model).__name__ if inner_model else "None"
-                
-                # Format inner_model_id properly for f-string
-                inner_id_str = f"0x{inner_model_id:x}" if inner_model_id is not None else "None"
-                
-                logger.mgpu_mm_log(f"[OBJECT_CHAIN_SET] ModelPatcher via patcher: mp_id=0x{mp_id:x}, inner_model_id={inner_id_str}, inner_model_type={inner_model_name}")
-                
-                # FIX: Store flag on ModelPatcher itself
-                mp._mgpu_unload_distorch_model = unload_distorch_model
-                logger.mgpu_mm_log(f"[FLAG_SET_LOCATION] Set on ModelPatcher (mp_id=0x{mp_id:x}): mp._mgpu_unload_distorch_model = {unload_distorch_model}")
-                
-                # Also set on inner model for backwards compatibility
-                if inner_model:
-                    inner_model._mgpu_unload_distorch_model = unload_distorch_model
-                    logger.mgpu_mm_log(f"[FLAG_SET_COMPAT] Also set on inner model (inner_model_id=0x{inner_model_id:x}) for compatibility")
-
-            if unload_distorch_model:
-                logger.mgpu_mm_log("[FLAG_TRIGGER] unload_distorch_model=True, triggering full system cleanup")
-                force_full_system_cleanup(reason="policy_every_load", force=True)
-
-            return out
-
-    return NodeOverrideDisTorchSafetensorV2
-
-
-def override_class_with_distorch_safetensor_v2_clip(cls):
-    """DisTorch 2.0 wrapper for safetensor CLIP models"""
-
-    class NodeOverrideDisTorchSafetensorV2Clip(cls):
-        @classmethod
-        def INPUT_TYPES(s):
-            inputs = copy.deepcopy(cls.INPUT_TYPES())
-            devices = get_device_list()
-            default_device = devices[1] if len(devices) > 1 else devices[0]
-
-            inputs["optional"] = inputs.get("optional", {})
-            inputs["optional"]["device"] = (devices, {"default": default_device})  # Changed from compute_device
-            inputs["optional"]["virtual_vram_gb"] = ("FLOAT", {"default": 4.0, "min": 0.0, "max": 128.0, "step": 0.1})
-            inputs["optional"]["donor_device"] = (devices, {"default": "cpu"})
-            inputs["optional"]["expert_mode_allocations"] = ("STRING", {"multiline": False, "default": ""})
-            inputs["optional"]["keep_loaded"] = ("BOOLEAN", {"default": True})
-            return inputs
-
-        CATEGORY = "multigpu/distorch_2"
-        FUNCTION = "override"
-        TITLE = f"{cls.TITLE if hasattr(cls, 'TITLE') else cls.__name__} (DisTorch2)"
-
-        @classmethod
-        def IS_CHANGED(s, *args, device=None, virtual_vram_gb=4.0,  # Changed from compute_device
-                       donor_device="cpu", expert_mode_allocations="", keep_loaded=True, **kwargs):
-            # Create a hash of our specific settings
-            settings_str = f"{device}{virtual_vram_gb}{donor_device}{expert_mode_allocations}{keep_loaded}"  # Changed from compute_device
-            current_hash = hashlib.sha256(settings_str.encode()).hexdigest()
-
-            if not hasattr(cls, '_last_hash'):
-                cls._last_hash = current_hash
-                logger.mgpu_mm_log(f"IS_CHANGED first call: {current_hash[:8]}")
-            elif cls._last_hash != current_hash:
-                cls._last_hash = current_hash
-                logger.mgpu_mm_log(f"IS_CHANGED CHANGED: {current_hash[:8]} ← settings changed")
-            return current_hash
-
-        def override(self, *args, device=None, virtual_vram_gb=4.0,  # Changed from compute_device
-                     donor_device="cpu", expert_mode_allocations="", keep_loaded=True, **kwargs):
-
-            unload_distorch_model = not keep_loaded
-
-            from . import set_current_text_encoder_device  # Use text encoder device setter
-            if device is not None:
-                set_current_text_encoder_device(device)
-
-            kwargs['device'] = 'default'  # Hardcode device setting like in standard clip wrapper
-
-            # Register our patched ModelPatcher
-            register_patched_safetensor_modelpatcher()
-
-            # Call original function
-            fn = getattr(super(), cls.FUNCTION)
-
-            # Call the main function once
-            out = fn(*args, **kwargs)
-
-            logger.mgpu_mm_log(f"[FLAG_SET_START] Setting '_mgpu_unload_distorch_model' to: {unload_distorch_model} (keep_loaded={keep_loaded})")
-
-            # DIAGNOSTIC: Log full object chain at SET time
-            if hasattr(out[0], 'model'):
-                mp = out[0]  # This is the ModelPatcher
-                mp_id = id(mp)
-                inner_model = getattr(mp, 'model', None)
-                inner_model_id = id(inner_model) if inner_model else None
-                inner_model_name = type(inner_model).__name__ if inner_model else "None"
-                
-                # Format inner_model_id properly for f-string
-                inner_id_str = f"0x{inner_model_id:x}" if inner_model_id is not None else "None"
-                
-                logger.mgpu_mm_log(f"[OBJECT_CHAIN_SET] CLIP ModelPatcher: mp_id=0x{mp_id:x}, inner_model_id={inner_id_str}, inner_model_type={inner_model_name}")
-                
-                # FIX: Store flag on ModelPatcher itself
-                mp._mgpu_unload_distorch_model = unload_distorch_model
-                logger.mgpu_mm_log(f"[FLAG_SET_LOCATION] Set on ModelPatcher (mp_id=0x{mp_id:x}): mp._mgpu_unload_distorch_model = {unload_distorch_model}")
-                
-                # Also set on inner model for backwards compatibility
-                if inner_model:
-                    inner_model._mgpu_unload_distorch_model = unload_distorch_model
-                    logger.mgpu_mm_log(f"[FLAG_SET_COMPAT] Also set on inner model (inner_model_id=0x{inner_model_id:x}) for compatibility")
-                    
-            elif hasattr(out[0], 'patcher') and hasattr(out[0].patcher, 'model'):
-                mp = out[0].patcher  # This is the ModelPatcher
-                mp_id = id(mp)
-                inner_model = getattr(mp, 'model', None)
-                inner_model_id = id(inner_model) if inner_model else None
-                inner_model_name = type(inner_model).__name__ if inner_model else "None"
-                
-                # Format inner_model_id properly for f-string
-                inner_id_str = f"0x{inner_model_id:x}" if inner_model_id is not None else "None"
-                
-                logger.mgpu_mm_log(f"[OBJECT_CHAIN_SET] CLIP ModelPatcher via patcher: mp_id=0x{mp_id:x}, inner_model_id={inner_id_str}, inner_model_type={inner_model_name}")
-                
-                # FIX: Store flag on ModelPatcher itself
-                mp._mgpu_unload_distorch_model = unload_distorch_model
-                logger.mgpu_mm_log(f"[FLAG_SET_LOCATION] Set on ModelPatcher (mp_id=0x{mp_id:x}): mp._mgpu_unload_distorch_model = {unload_distorch_model}")
-                
-                # Also set on inner model for backwards compatibility
-                if inner_model:
-                    inner_model._mgpu_unload_distorch_model = unload_distorch_model
-                    logger.mgpu_mm_log(f"[FLAG_SET_COMPAT] Also set on inner model (inner_model_id=0x{inner_model_id:x}) for compatibility")
-
-            vram_string = ""
-            if virtual_vram_gb > 0:
-                vram_string = f"{device};{virtual_vram_gb};{donor_device}"
-            elif expert_mode_allocations:
-                vram_string = device
-
-            full_allocation = f"{expert_mode_allocations}#{vram_string}" if expert_mode_allocations or vram_string else ""
-
-            logger.info(f"[MultiGPU DisTorch V2] Full allocation string: {full_allocation}")
-
-            # Store allocation AFTER loading for next time
-            model_to_check = None
-            if hasattr(out[0], 'model'):
-                model_to_check = out[0]
-            elif hasattr(out[0], 'patcher') and hasattr(out[0].patcher, 'model'):
-                model_to_check = out[0].patcher
-
-            if model_to_check:
-                model_hash = create_safetensor_model_hash(model_to_check, "override_store")
-                settings_str = f"{device}{virtual_vram_gb}{donor_device}{expert_mode_allocations}"
-                settings_hash = hashlib.sha256(settings_str.encode()).hexdigest()
-                
-                # Store allocation for next time
-                safetensor_allocation_store[model_hash] = full_allocation
-                safetensor_settings_store[model_hash] = settings_hash
-
-            if unload_distorch_model:
-                logger.mgpu_mm_log("[FLAG_TRIGGER] unload_distorch_model=True, triggering full system cleanup")
-                force_full_system_cleanup(reason="policy_every_load", force=True)
-
-            return out
-
-    return NodeOverrideDisTorchSafetensorV2Clip
-
-def override_class_with_distorch_safetensor_v2_clip_no_device(cls):
-    """DisTorch 2.0 wrapper for safetensor CLIP models"""
-
-    class NodeOverrideDisTorchSafetensorV2ClipNoDevice(cls):
-        @classmethod
-        def INPUT_TYPES(s):
-            inputs = copy.deepcopy(cls.INPUT_TYPES())
-            devices = get_device_list()
-            default_device = devices[1] if len(devices) > 1 else devices[0]
-
-            inputs["optional"] = inputs.get("optional", {})
-            inputs["optional"]["device"] = (devices, {"default": default_device})  # Changed from compute_device
-            inputs["optional"]["virtual_vram_gb"] = ("FLOAT", {"default": 4.0, "min": 0.0, "max": 128.0, "step": 0.1})
-            inputs["optional"]["donor_device"] = (devices, {"default": "cpu"})
-            inputs["optional"]["expert_mode_allocations"] = ("STRING", {"multiline": False, "default": ""})
-            inputs["optional"]["keep_loaded"] = ("BOOLEAN", {"default": True})
-            return inputs
-
-        CATEGORY = "multigpu/distorch_2"
-        FUNCTION = "override"
-        TITLE = f"{cls.TITLE if hasattr(cls, 'TITLE') else cls.__name__} (DisTorch2)"
-
-        @classmethod
-        def IS_CHANGED(s, *args, device=None, virtual_vram_gb=4.0,  # Changed from compute_device
-                       donor_device="cpu", expert_mode_allocations="", keep_loaded=True, **kwargs):
-            # Create a hash of our specific settings
-            settings_str = f"{device}{virtual_vram_gb}{donor_device}{expert_mode_allocations}{keep_loaded}"  # Changed from compute_device
-            current_hash = hashlib.sha256(settings_str.encode()).hexdigest()
-
-            if not hasattr(cls, '_last_hash'):
-                cls._last_hash = current_hash
-                logger.mgpu_mm_log(f"IS_CHANGED first call: {current_hash[:8]}")
-            elif cls._last_hash != current_hash:
-                cls._last_hash = current_hash
-                logger.mgpu_mm_log(f"IS_CHANGED CHANGED: {current_hash[:8]} ← settings changed")
-            return current_hash
-        def override(self, *args, device=None, virtual_vram_gb=4.0,  # Changed from compute_device
-                     donor_device="cpu", expert_mode_allocations="", keep_loaded=True, **kwargs):
-
-            unload_distorch_model = not keep_loaded
-
-            from . import set_current_text_encoder_device  # Use text encoder device setter
-            if device is not None:
-                set_current_text_encoder_device(device)
-
-            # Register our patched ModelPatcher
-            register_patched_safetensor_modelpatcher()
-
-            # Call original function
-            fn = getattr(super(), cls.FUNCTION)
-
-            # Call the main function once
-            out = fn(*args, **kwargs)
-
-            logger.mgpu_mm_log(f"[FLAG_SET_START] Setting '_mgpu_unload_distorch_model' to: {unload_distorch_model} (keep_loaded={keep_loaded})")
-
-            # DIAGNOSTIC: Log full object chain at SET time
-            if hasattr(out[0], 'model'):
-                mp = out[0]  # This is the ModelPatcher
-                mp_id = id(mp)
-                inner_model = getattr(mp, 'model', None)
-                inner_model_id = id(inner_model) if inner_model else None
-                inner_model_name = type(inner_model).__name__ if inner_model else "None"
-                
-                # Format inner_model_id properly for f-string
-                inner_id_str = f"0x{inner_model_id:x}" if inner_model_id is not None else "None"
-                
-                logger.mgpu_mm_log(f"[OBJECT_CHAIN_SET] CLIP_NoDevice ModelPatcher: mp_id=0x{mp_id:x}, inner_model_id={inner_id_str}, inner_model_type={inner_model_name}")
-                
-                # FIX: Store flag on ModelPatcher itself
-                mp._mgpu_unload_distorch_model = unload_distorch_model
-                logger.mgpu_mm_log(f"[FLAG_SET_LOCATION] Set on ModelPatcher (mp_id=0x{mp_id:x}): mp._mgpu_unload_distorch_model = {unload_distorch_model}")
-                
-                # Also set on inner model for backwards compatibility
-                if inner_model:
-                    inner_model._mgpu_unload_distorch_model = unload_distorch_model
-                    logger.mgpu_mm_log(f"[FLAG_SET_COMPAT] Also set on inner model (inner_model_id=0x{inner_model_id:x}) for compatibility")
-                    
-            elif hasattr(out[0], 'patcher') and hasattr(out[0].patcher, 'model'):
-                mp = out[0].patcher  # This is the ModelPatcher
-                mp_id = id(mp)
-                inner_model = getattr(mp, 'model', None)
-                inner_model_id = id(inner_model) if inner_model else None
-                inner_model_name = type(inner_model).__name__ if inner_model else "None"
-                
-                # Format inner_model_id properly for f-string
-                inner_id_str = f"0x{inner_model_id:x}" if inner_model_id is not None else "None"
-                
-                logger.mgpu_mm_log(f"[OBJECT_CHAIN_SET] CLIP_NoDevice ModelPatcher via patcher: mp_id=0x{mp_id:x}, inner_model_id={inner_id_str}, inner_model_type={inner_model_name}")
-                
-                # FIX: Store flag on ModelPatcher itself
-                mp._mgpu_unload_distorch_model = unload_distorch_model
-                logger.mgpu_mm_log(f"[FLAG_SET_LOCATION] Set on ModelPatcher (mp_id=0x{mp_id:x}): mp._mgpu_unload_distorch_model = {unload_distorch_model}")
-                
-                # Also set on inner model for backwards compatibility
-                if inner_model:
-                    inner_model._mgpu_unload_distorch_model = unload_distorch_model
-                    logger.mgpu_mm_log(f"[FLAG_SET_COMPAT] Also set on inner model (inner_model_id=0x{inner_model_id:x}) for compatibility")
-
-            vram_string = ""
-            if virtual_vram_gb > 0:
-                vram_string = f"{device};{virtual_vram_gb};{donor_device}"
-            elif expert_mode_allocations:
-                vram_string = device
-
-            full_allocation = f"{expert_mode_allocations}#{vram_string}" if expert_mode_allocations or vram_string else ""
-
-            logger.info(f"[MultiGPU DisTorch V2] Full allocation string: {full_allocation}")
-
-            # Store allocation AFTER loading for next time
-            model_to_check = None
-            if hasattr(out[0], 'model'):
-                model_to_check = out[0]
-            elif hasattr(out[0], 'patcher') and hasattr(out[0].patcher, 'model'):
-                model_to_check = out[0].patcher
-
-            if model_to_check:
-                model_hash = create_safetensor_model_hash(model_to_check, "override_store")
-                settings_str = f"{device}{virtual_vram_gb}{donor_device}{expert_mode_allocations}"
-                settings_hash = hashlib.sha256(settings_str.encode()).hexdigest()
-                
-                # Store allocation for next time
-                safetensor_allocation_store[model_hash] = full_allocation
-                safetensor_settings_store[model_hash] = settings_hash
-
-            if unload_distorch_model:
-                logger.mgpu_mm_log("[FLAG_TRIGGER] unload_distorch_model=True, triggering full system cleanup")
-                force_full_system_cleanup(reason="policy_every_load", force=True)
-
-            return out
-
-    return NodeOverrideDisTorchSafetensorV2ClipNoDevice
+# NOTE: All wrapper functions have been moved to wrappers.py for better organization.
+# This file (distorch_2.py) now contains ONLY backend logic:
+# - register_patched_safetensor_modelpatcher()
+# - analyze_safetensor_loading() and analyze_safetensor_loading_clip()
+# - calculate_safetensor_vvram_allocation()
+# - Allocation stores and model hash functions
